Editor prerequisites [part 2] (#889)
