+++ lib/filesystem/Filesystem.cpp
@@ -225,6 +225,21 @@ void CResourceHandler::addFilesystem(const std::string & parent, const std::stri
 	knownLoaders[identifier] = loader;
 }
 
+bool CResourceHandler::removeFilesystem(const std::string & parent, const std::string & identifier)
+{
+	if(knownLoaders.count(identifier) == 0)
+		return false;
+	
+	if(knownLoaders.count(parent) == 0)
+		return false;
+	
+	auto list = dynamic_cast<CFilesystemList *>(knownLoaders.at(parent));
+	assert(list);
+	list->removeLoader(knownLoaders[identifier]);
+	knownLoaders.erase(identifier);
+	return true;
+}
+
 ISimpleResourceLoader * CResourceHandler::createFileSystem(const std::string & prefix, const JsonNode &fsConfig)
 {
 	CFilesystemGenerator generator(prefix);
