+++ lib/mapObjects/CObjectHandler.cpp
@@ -390,6 +390,11 @@ void CGObjectInstance::afterAddToMap(CMap * map)
 	//nothing here
 }
 
+void CGObjectInstance::afterRemoveFromMap(CMap * map)
+{
+	//nothing here
+}
+
 void CGObjectInstance::serializeJsonOptions(JsonSerializeFormat & handler)
 {
 	//nothing here
