+++ lib/mapObjects/MiscObjects.cpp
@@ -1425,6 +1425,9 @@ void CGArtifact::blockingDialogAnswered(const CGHeroInstance *hero, ui32 answer)
 
 void CGArtifact::afterAddToMap(CMap * map)
 {
+	//Artifacts from map objects are never removed
+	//FIXME: This should be revertible in map editor
+
 	if(ID == Obj::SPELL_SCROLL && storedArtifact && storedArtifact->id.getNum() < 0)
         map->addNewArtifactInstance(storedArtifact);
 }
