+++ lib/mapObjects/MiscObjects.h
@@ -445,7 +445,7 @@ public:
 	}
 };
 
-class CGShipyard : public CGObjectInstance, public IShipyard
+class DLL_LINKAGE CGShipyard : public CGObjectInstance, public IShipyard
 {
 public:
 	void getOutOffsets(std::vector<int3> &offsets) const override; //offsets to obj pos when we boat can be placed
