+++ lib/mapping/CMap.h
@@ -361,9 +361,14 @@ public:
 	void eraseArtifactInstance(CArtifactInstance * art);
 
 	void addNewQuestInstance(CQuest * quest);
+	void removeQuestInstance(CQuest * quest);
 
+	void setUniqueInstanceName(CGObjectInstance * obj);
 	///Use only this method when creating new map object instances
 	void addNewObject(CGObjectInstance * obj);
+	void moveObject(CGObjectInstance * obj, const int3 & dst);
+	void removeObject(CGObjectInstance * obj);
+
 
 	/// Gets object of specified type on requested position
 	const CGObjectInstance * getObjectiveObjectFrom(int3 pos, Obj::EObj type);
@@ -408,6 +413,7 @@ public:
 private:
 	/// a 3-dimensional array of terrain tiles, access is as follows: x, y, level. where level=1 is underground
 	TerrainTile*** terrain;
+	si32 uidCounter; //TODO: initialize when loading an old map
 
 public:
 	template <typename Handler>
