+++ lib/rmg/CRmgTemplateStorage.cpp
@@ -32,6 +32,8 @@ void CRmgTemplateStorage::loadObject(std::string scope, std::string name, const 
 		templates[fullKey].setId(name);
 		templates[fullKey].serializeJson(handler);
 		templates[fullKey].validate();
+
+		templatesByName[name] = templates[fullKey];
 	}
 	catch(const std::exception & e)
 	{
@@ -51,10 +53,18 @@ std::vector<JsonNode> CRmgTemplateStorage::loadLegacyData(size_t dataSize)
 	//it would be cool to load old rmg.txt files
 }
 
-const CRmgTemplate * CRmgTemplateStorage::getTemplate(const std::string & templateName) const
+const CRmgTemplate * CRmgTemplateStorage::getTemplate(const std::string & templateFullId) const
 {
-	auto iter = templates.find(templateName);
+	auto iter = templates.find(templateFullId);
 	if(iter==templates.end())
+		return nullptr;
+	return &iter->second;
+}
+
+const CRmgTemplate * CRmgTemplateStorage::getTemplateByName(const std::string & templateName) const
+{
+	auto iter = templatesByName.find(templateName);
+	if(iter == templatesByName.end())
 		return nullptr;
 	return &iter->second;
 }
